Extend underline sliding behavior for menu options in navigation #847
